backend/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
twilio>=9.6.5
gunicorn==21.2.0
africas-talking==3.2.0
openai==0.27.8
werkzeug==3.0.1